Prod/Core/requirements.txt
aniso8601==9.0.1
atomicwrites==1.4.0
attrs==20.3.0
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
Flask==2.0.2
Flask-RESTful==0.3.9
idna==3.3
importlib-metadata==3.4.0
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
packaging==20.9
Pillow==8.4.0
pip==21.3.1
pluggy==0.13.1
psycopg2==2.9.2
py==1.10.0
pyasn1==0.4.8
PyJWT==2.3.0
pyparsing==2.4.7
pytest==6.2.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.26.0
rsa==4.7.2
schedule==1.1.0
setuptools==47.1.0
six==1.16.0
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.7
Werkzeug==2.0.2
zipp==3.4.0
